Separate header files from source files
