+++ .gitignore
@@ -49,7 +49,6 @@ include/divisible/*
 
 build/**/*
 
-include/*
 lib/*
 bin/*
 test/test_runner
