+++ include/app/parallel_rdp_wrapper.h
