+++ src/app/CMakeLists.txt
@@ -1,7 +1,7 @@
 include_directories(${SDL2_INCLUDE_DIRS})
 add_library(app
-    app.h app.cpp
-    parallel_rdp_wrapper.h parallel_rdp_wrapper.cpp
+    app.cpp
+    parallel_rdp_wrapper.cpp
 )
 target_precompile_headers(app PRIVATE ${PCH_FILES})
 target_link_libraries(app PRIVATE
