+++ src/app/app.cpp
@@ -1,7 +1,7 @@
-#include "app.h"
+#include "app/app.h"
+#include "app/parallel_rdp_wrapper.h"
 #include "memory/memory.h"
 #include "n64_system/n64_system.h"
-#include "parallel_rdp_wrapper.h"
 #include <SDL.h>
 #include <SDL_vulkan.h>
