+++ src/cpu/CMakeLists.txt
@@ -1,7 +1,7 @@
 add_library(cpu
-    cpu.h cpu.cpp
-    cop0.h cop0.cpp
-    cop1.h cop1.cpp
+    cpu.cpp
+    cop0.cpp
+    cop1.cpp
     instruction.cpp
     cpu_instruction_impl.cpp
     fpu_instruction_impl.cpp
