+++ src/cpu/cpu_instruction_impl.cpp
@@ -1,4 +1,4 @@
-#include "cpu/cpu_instruction_impl.h"
+#include "cpu_instruction_impl.h"
 #include "cpu/cpu.h"
 #include "memory/bus.h"
 #include "mmu/mmu.h"
