+++ src/memory/bus.cpp
@@ -1,6 +1,6 @@
-﻿#include "bus.h"
-#include "memory.h"
-#include "memory_map.h"
+﻿#include "memory/bus.h"
+#include "memory/memory.h"
+#include "memory/memory_map.h"
 #include "mmio/ai.h"
 #include "mmio/mi.h"
 #include "mmio/pi.h"
