+++ src/memory/memory.cpp
@@ -1,5 +1,5 @@
-﻿#include "memory.h"
-#include "memory_map.h"
+﻿#include "memory/memory.h"
+#include "memory/memory_map.h"
 #include <iostream>
 
 namespace N64 {
