+++ src/mmu/CMakeLists.txt
@@ -1,4 +1,4 @@
 add_library(mmu
-    mmu.cpp mmu.h
-    tlb.cpp tlb.h)
+    mmu.cpp
+    tlb.cpp)
 target_precompile_headers(mmu PRIVATE ${PCH_FILES})
