+++ src/n64_system/n64_system.cpp
@@ -1,6 +1,5 @@
-﻿#include "n64_system.h"
+﻿#include "n64_system/n64_system.h"
 #include "app/parallel_rdp_wrapper.h"
-#include "config.h"
 #include "memory/bus.h"
 #include "memory/memory.h"
 #include "mmio/ai.h"
@@ -9,6 +8,7 @@
 #include "mmio/si.h"
 #include "mmio/vi.h"
 #include "mmu/tlb.h"
+#include "n64_system/config.h"
 #include "n64_system/interrupt.h"
 #include "n64_system/scheduler.h"
 #include "rcp/rsp.h"
