+++ src/utils/CMakeLists.txt
@@ -1,5 +1,5 @@
 add_library(utils
-    utils.cpp utils.h)
+    utils.cpp)
 target_precompile_headers(utils PRIVATE ${PCH_FILES})
 target_link_libraries(utils PRIVATE
     cpu
